+++ functional_tests/example_cre.R
@@ -1,16 +1,9 @@
-set.seed(2022)
-
-# Set Experiment Parameter
-n_rules <- 2
-sample_size <- 2000
-effect_size <- 2
-confounding <- "no"
-ite_estimator <- "aipw"
-pfer <- n_rules/(effect_size+1)
+set.seed(2023)
+n_rules <- 3
 
 # Set Method and Hyper Parameters
 method_params <- list(ratio_dis = 0.5,
-                      ite_method = ite_estimator,
+                      ite_method = "aipw",
                       learner_ps = "SL.xgboost",
                       learner_y = "SL.xgboost")
 
@@ -21,14 +14,14 @@ hyper_params <- list(intervention_vars = NULL,
                      max_rules = 50,
                      max_depth = 2,
                      t_decay = 0.025,
-                     t_ext = 0.01,
+                     t_ext = 0.025,
                      t_corr = 1,
                      t_pvalue = 0.05,
-                     stability_selection = TRUE,
+                     stability_selection = "vanilla",
                      cutoff = 0.8,
-                     pfer = pfer,
-                     B = 20,
-                     subsample = 0.2)
+                     pfer = 1,
+                     B = 50,
+                     subsample = 0.05)
 
 # Set Ground Truth
 {
@@ -52,14 +45,14 @@ hyper_params <- list(intervention_vars = NULL,
 }
 
 # Generate Dataset
-dataset <- generate_cre_dataset(n = sample_size,
+dataset <- generate_cre_dataset(n = 5000,
                                 rho = 0,
                                 p = 10,
-                                effect_size = effect_size,
+                                effect_size = 5,
                                 n_rules = n_rules,
                                 binary_covariates = TRUE,
                                 binary_outcome = FALSE,
-                                confounding = confounding)
+                                confounding = "no")
 y <- dataset[["y"]]
 z <- dataset[["z"]]
 X <- dataset[["X"]]